Move remove `webauthn` namespace from passkeys controller
The goal is to stop hardcoding the controller to "devise/webauthn/passkeys"
to allow the user to override the controller.
Instead, `controllers[:passkeys]` should be used.

But `controllers[:passkeys]` defaults to `devise/passkeys` instead of
`devise/webauthn/passkeys` as can be seen here: https://github.com/heartcombo/devise/blob/cf93de390a29654620fdf7ac07b4794eb95171d0/lib/devise/mapping.rb#L131-L136
